Old enough versions of g++ get upset if you have integer literals that are too large, and seem to need INT64_C() wrapped around them even though newer C++ standrards would not require this. So put in INT64_C() in a bunch of places. Put in some more braces in initialisers to get them more pedantically correct. Arrange that the soft float library can use inline functions - noting that what it says about that in its documentation applied to older versions of gcc but if you use "extern inline" now you will find that compilation FAILS with a load of multiply defined functions. I hope I have tested the changes here from the 32-bit version of Fedora 9 up to Ubuntu 15.10 and Windows 10.
git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@3608 2bfe0521-f11c-4a00-b80e-6202646ff360

+++ csl/cslbase/cslread.cpp
@@ -121,8 +121,8 @@ void validate_string_fn(LispObject s, const char *file, int line)
                 fprintf(stderr, "\n+++ Bad string at %s %d\n", file, line);
                 fprintf(stderr, "Header = %" PRIxPTR "\n", vechdr(s));
                 fprintf(stderr, "length = %d bytelength = %d\n",
-                    length_of_header(vechdr(s)),
-                    length_of_byteheader(vechdr(s)));
+                    (int)length_of_header(vechdr(s)),
+                    (int)length_of_byteheader(vechdr(s)));
                 fprintf(stderr, "messed at len:%d len1:%d [%x]\n",
                         len, len1, celt(s, len-CELL));
                 for (i=0; i<len1; i++)
